Use total counts instead of mean for distribution.
The mean value has the sideffect that we would have to worry about
elements (transcripts, exons, etc) with all 0s in all bins. Should we
keep them or not? To avoid this we now use the sum and the density is
calculated by dividing with the total number of reads. Also the output
now has all the elements instead of just the distribution enabling to
figure out outliers manually. Obviously the plotting scripts have been
updated accordingly to support the new output format.